backend/requirements.txt
# HomeGuardian AI — Python Dependencies
# All versions pinned for reproducibility

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
# SQLite is built into Python — no package needed

# MQTT
paho-mqtt==1.6.1

# Computer Vision
opencv-python==4.8.1.78
ultralytics==8.0.227

# Machine Learning
scikit-learn==1.3.2
numpy==1.26.2
scipy==1.11.4

# AI Narrative Generation
anthropic==0.7.8

# Push Notifications
firebase-admin==6.2.0

# Rate Limiting
slowapi==0.1.9

# WebSocket
websockets==12.0

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0
